% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptCropMask.R
\name{ScriptCropMask}
\alias{ScriptCropMask}
\title{This package performs cropping of raster files}
\usage{
ScriptCropMask(shapefile, Directory)
}
\arguments{
\item{shapefile}{shapefile local}

\item{Directory}{Location where the files will be saved}
}
\description{
This package performs cropping of raster files
}
